requirements.txt
appdirs==1.4.2
Django==1.10.6
django-crispy-forms==1.6.1
django-debug-toolbar==1.7
django-registration==2.2
django-webpack-loader==0.4.1
djangorestframework==3.6.2
flake8==3.3.0
jedi==0.10.0
mccabe==0.6.1
oauthlib==2.0.1
packaging==16.8
psycopg2==2.7
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
requests==2.13.0
requests-oauthlib==0.8.0
rope-py3k==0.9.4.post1
sh==1.12.10
six==1.10.0
sqlparse==0.2.3
